Change nat_int_t back to long long int
I wanted to ensure it was 64-bit, but it breaks on systems where
long long int and int64_t are the same type.

+++ include/natalie/types.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
-#include <stdint.h>
+#include <limits.h>
 
-#define NAT_INT_MIN INT64_MIN
-#define NAT_INT_MAX INT64_MAX
+#define NAT_INT_MIN LLONG_MIN
+#define NAT_INT_MAX LLONG_MAX
 
 namespace Natalie {
 
-using nat_int_t = int64_t;
+using nat_int_t = long long;
 
 }
